+++ test/public/types/readme-generated/32-1323-1344.ts
@@ -12,7 +12,7 @@ const users = Table.make("users", {
 })
 
 const insertUser = Query.insert(users, {
-  id: "11111111-1111-1111-1111-111111111111",
+  id: "11111111-1111-4111-8111-111111111111",
   email: "alice@example.com",
   visits: 1
 })